// frontend/layout/lib/types.ts
export enum FlexDirection {
    Row = "row",
    Column = "column",
}
// ...
export interface LayoutNode {
    id: string;
    data?: TabLayoutData;
    children?: LayoutNode[];
    flexDirection: FlexDirection;
    size: number;
}